fix: Prevent db from being created by default
We use a single db postgres for review apps so KISS, this commit creates
an app, allocates ips, attaches the existing db to the $INPUT_POSTGRES
variable value and finally deploys the config.

A future update should allow creating the db automatically, or
deprecating this functionality in the docs and options.